add images

+++ controllers/AuthorsController.php
@@ -2,6 +2,7 @@
 
 namespace app\controllers;
 
+use app\models\MagazinesAuthors;
 use Yii;
 use app\models\Authors;
 use app\models\AuthorsSearch;
@@ -102,11 +103,16 @@ class AuthorsController extends Controller
      * @return mixed
      * @throws NotFoundHttpException if the model cannot be found
      */
-    public function actionDelete($id)
+    public function actionDelete(int $id)
     {
-        $this->findModel($id)->delete();
+        $relations = MagazinesAuthors::getRelations($id);
+        if (count($relations) < 1) {
+            $this->findModel($id)->delete();
 
-        return $this->redirect(['index']);
+            return $this->redirect(['index']);
+        } else {
+            return false;
+        }
     }
 
     /**
